Add problem generator, put problemgen + check_plan as fcts in utils.py
Added a problem generator. Put both plan checker and problem generator as functions in
utils.py (problemgen and check_plan). Run utils.py as a main file to test them,
producing a plan in 'problem' (also printing it out) and checking the plan from
plan5.txt, with the output in dict 'o'.
Note that plan5.txt was generated mostly by ChatGPT, after correcting errors detected by
check_plan.
Should I enforce that putting a block on a pile preserves the pile's position in plan?
And that putting a block on table makes it the last pile in plan?